common device transforms
