+++ annet/adapters/netbox/v24/storage.py
@@ -2,12 +2,14 @@ from logging import getLogger
 from typing import Optional, List
 
 from annet.adapters.netbox import models, common_models
+from annet.adapters.netbox.common.manufacturer import (
+    is_supported, get_hw, get_breed,
+)
+from annet.adapters.netbox.common.storage_opts import NetboxStorageOpts
 from annet.adapters.netbox.query import NetboxQuery
-from annet.annlib.netdev.views.hardware import HardwareView
 from annet.storage import Storage
 from . import api_models
 from .client import NetboxV24
-from ..common.storage_opts import NetboxStorageOpts
 
 logger = getLogger(__name__)
 
@@ -37,6 +39,9 @@ def extend_label(
 
 
 def extend_device(device: api_models.Device) -> models.Device:
+    manufacturer = device.device_type.manufacturer.name
+    model = device.device_type.model
+
     return models.Device(
         id=device.id,
         name=device.name,
@@ -62,8 +67,8 @@ def extend_device(device: api_models.Device) -> models.Device:
 
         fqdn=device.name,
         hostname=device.name,
-        hw=get_hw(device),
-        breed=get_breed(device),
+        hw=get_hw(manufacturer, model),
+        breed=get_breed(manufacturer, model),
         interfaces=[],
     )
 
@@ -128,7 +133,8 @@ class NetboxStorageV24(Storage):
             for device in self.netbox.all_devices(
                 name__ic=query.globs,
             ).results
-            if is_supported(device) and _match_query(query, device)
+            if is_supported(device.device_type.manufacturer.name)
+            if _match_query(query, device)
         }
         if device_ids:
             interfaces = self._load_interfaces(list(device_ids))
@@ -169,67 +175,3 @@ def _match_query(query: NetboxQuery, device_data: api_models.Device) -> bool:
         if subquery.strip() in device_data.name:
             return True
     return False
-
-
-def get_hw(device: api_models.Device):
-    manufacturer = device.device_type.manufacturer.name
-    model_name = device.device_type.model
-    # by some reason Netbox calls Mellanox SN as MSN, so we fix them here
-    if manufacturer == "Mellanox" and model_name.startswith("MSN"):
-        model_name = model_name.replace("MSN", "SN", 1)
-    hw = _vendor_to_hw(manufacturer + " " + model_name)
-    if not hw:
-        raise ValueError(f"unsupported manufacturer {manufacturer}")
-    return hw
-
-
-def get_breed(device: api_models.Device):
-    manufacturer = device.device_type.manufacturer.name
-    model_name = device.device_type.model
-    if manufacturer == "Huawei" and model_name.startswith("CE"):
-        return "vrp85"
-    elif manufacturer == "Huawei" and model_name.startswith("NE"):
-        return "vrp85"
-    elif manufacturer == "Huawei":
-        return "vrp55"
-    elif manufacturer == "Mellanox":
-        return "cuml2"
-    elif manufacturer == "Juniper":
-        return "jun10"
-    elif manufacturer == "Cisco":
-        return "ios12"
-    elif manufacturer == "Adva":
-        return "adva8"
-    elif manufacturer == "Arista":
-        return "eos4"
-    raise ValueError(f"unsupported manufacturer {manufacturer}")
-
-
-def is_supported(device: api_models.Device) -> bool:
-    manufacturer = device.device_type.manufacturer.name
-    if manufacturer not in (
-            "Huawei", "Mellanox", "Juniper", "Cisco", "Adva", "Arista",
-    ):
-        logger.warning("Unsupported manufacturer `%s`", manufacturer)
-        return False
-    return True
-
-
-def _vendor_to_hw(vendor):
-    hw = HardwareView(
-        {
-            "cisco": "Cisco",
-            "catalyst": "Cisco Catalyst",
-            "nexus": "Cisco Nexus",
-            "huawei": "Huawei",
-            "juniper": "Juniper",
-            "arista": "Arista",
-            "pc": "PC",
-            "nokia": "Nokia",
-            "aruba": "Aruba",
-            "routeros": "RouterOS",
-            "ribbon": "Ribbon",
-        }.get(vendor.lower(), vendor),
-        None,
-    )
-    return hw
